cosmetic: dns: Move `is_iface_valid_for_dns` to top level
The `is_iface_valid_for_dns` is been used in `dns.rs`, hence it should
not be stored in `nm/dns.rs`.
